Added direction for 3d sprites. Can be changed in Mapper. 0..5 standart dirs, any other value is absolute angle in degrees.
Some other minor changes and fixes.

git-svn-id: https://cvet.by:8443/svn/FOnlineSource@54 c2c6a35e-ad2b-4e1f-bb64-97b989be360b

Former-commit-id: b3a434393f64c1d9139a8c91fe9d98fe35370833

+++ Source/ProtoMap.h
@@ -24,8 +24,8 @@ public:
 	WORD ProtoId;
 	WORD MapX;
 	WORD MapY;
-	BYTE Dir;
-	// BYTE[3]
+	short Dir;
+	// BYTE[2]
 
 	DWORD LightColor;
 	BYTE LightDay;
